refactor(packages/eg-walker): complete OPERATION_TYPE import refactoring and remove legacy exports
